+++ app/src/main/java/com/insightapp/nocontrole/model/repository/LanctoRepository.kt
@@ -4,12 +4,13 @@ import androidx.lifecycle.LiveData
 import com.insightapp.nocontrole.model.dao.LanctoDao
 import com.insightapp.nocontrole.model.entity.Lancto
 
+/*Classe intermediária responsável por disponibilizar e/ou manutenção de dados,
+* através da biblioteca room ou retrofit*/
 class LanctoRepository (private val lanctoDao: LanctoDao, private val month: Int = 0) {
 
     val allLanctos: LiveData<List<Lancto>> = lanctoDao.getAll()
-    val totByDescByMonth: LiveData<Double> = lanctoDao.totByDescByMonth(month)
-    val totByRecByMonth: LiveData<Double> = lanctoDao.totByRecByMonth(month)
-
+    val totByDescByMonth: LiveData<Float> = lanctoDao.totByDescByMonth(month)
+    val totByRecByMonth: LiveData<Float> = lanctoDao.totByRecByMonth(month)
 
     suspend fun insert(lancto: Lancto) {
         lanctoDao.insert(lancto)
